requirements.txt
openai==1.12.0
python-dotenv==1.0.1
requests==2.31.0
huggingface-hub==0.20.3
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
